delete button for saved projects
